v1.0 | 1/31/20
Basic new TicTacToe game with a computer that generates a random integer and plays on that spot. The basic game does not allow any illegal moves. The computer moves fairly and the player is forced to move fairly too. The next versions will be the beginning of the arrival of the AI.