v1.1 | 2/5/20
This upload adds some code to the Brutus class. Brutus now can play on its own, but still is random. Once it gets its turn it wins the game and then exits. So, in the next iteration that will be fixed. Everything else about the game is the same. Only the Brutus class has changed.